# cmd/fmc-go-agents-server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "fmc-go-agents-server服务Api文档",
        "title": "fmc-go-agents-server服务Api文档",
        "termsOfService": "fmc-go-agents-server-最终解释权",
        "contact": {
            "name": "fmc-go-agents-server freedqo"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/session/creatSession": {
            "post": {
                "description": "创建会话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话管理"
                ],
                "summary": "创建会话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tokenid 用户登录令牌",
                        "name": "Tokenid",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "sessionm.CreatSessionReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sessionm.CreatSessionReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/webapp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sessionm.CreatSessionResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/session/deleteChatLogs": {
            "delete": {
                "description": "删除多条对话记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话管理"
                ],
                "summary": "删除多条对话记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tokenid 用户登录令牌",
                        "name": "Tokenid",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "sessionm.DeleteChatLogsReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sessionm.DeleteChatLogsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/webapp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sessionm.DeleteChatLogsResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/session/deleteSessions": {
            "delete": {
                "description": "删除多个用户对话",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话管理"
                ],
                "summary": "删除多个用户对话",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tokenid 用户登录令牌",
                        "name": "Tokenid",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "sessionm.DeleteSessionsReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sessionm.DeleteSessionsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/webapp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sessionm.DeleteSessionsResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/session/querySessionChatLogsByUser": {
            "post": {
                "description": "根据用户信息和提示词类型获取唯一会话记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话管理"
                ],
                "summary": "根据用户信息和提示词类型获取唯一会话记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tokenid 用户登录令牌",
                        "name": "Tokenid",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "sessionm.QuerySessionChatLogsByUserReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sessionm.QuerySessionChatLogsByUserReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/webapp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sessionm.QuerySessionChatLogsByUserResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/session/sessionChatLogList": {
            "post": {
                "description": "查询聊天内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话管理"
                ],
                "summary": "查询聊天内容",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tokenid 用户登录令牌",
                        "name": "Tokenid",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "sessionm.SessionChatLogListReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sessionm.SessionChatLogListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/webapp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sessionm.SessionChatLogListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/session/userSessionList": {
            "post": {
                "description": "查询用户对话列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "会话管理"
                ],
                "summary": "查询用户对话列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tokenid 用户登录令牌",
                        "name": "Tokenid",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求参数",
                        "name": "sessionm.UserSessionListReq",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/sessionm.UserSessionListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功响应",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/webapp.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/sessionm.UserSessionListResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "dbm.Page": {
            "type": "object",
            "properties": {
                "index": {
                    "description": "当前页码",
                    "type": "integer"
                },
                "size": {
                    "description": "每页大小",
                    "type": "integer"
                },
                "total": {
                    "description": "总条数",
                    "type": "integer"
                }
            }
        },
...
        "model.Ai_chat_logs": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "消息内容",
                    "type": "string"
                },
                "created_at": {
                    "description": "消息创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "记录唯一id",
                    "type": "string"
                },
                "order": {
                    "description": "消息顺序",
                    "type": "integer"
                },
                "role": {
                    "description": "消息角色",
                    "type": "string"
                },
                "session_id": {
                    "description": "会话唯一标识",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户ID",
                    "type": "string"
                }
            }
        },
...
        "sessionm.ChatLog": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "消息内容",
                    "type": "string"
                },
                "role": {
                    "description": "消息角色",
                    "type": "string"
                },
                "timestamp": {
                    "description": "消息时间戳",
                    "type": "integer"
                }
            }
        },
        "sessionm.CreatSessionReq": {
            "type": "object",
            "properties": {
                "userId": {
                    "type": "string"
                }
            }
        },
        "sessionm.CreatSessionResp": {
            "type": "object",
            "properties": {
                "sessionId": {
                    "type": "string"
                }
            }
        },
        "sessionm.DeleteChatLogsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "sessionm.DeleteChatLogsResp": {
            "type": "object"
        },
        "sessionm.DeleteSessionsReq": {
            "type": "object",
            "properties": {
                "sessionIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "sessionm.DeleteSessionsResp": {
            "type": "object"
        },
        "sessionm.QuerySessionChatLogsByUserReq": {
            "type": "object",
            "properties": {
                "promptType": {
                    "description": "提示词类型",
                    "type": "string"
                },
                "userId": {
                    "description": "用户id",
                    "type": "string"
                }
            }
        },
        "sessionm.QuerySessionChatLogsByUserResp": {
            "type": "object",
            "properties": {
                "chatLogs": {
                    "description": "聊天记录",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sessionm.ChatLog"
                    }
                },
                "promptType": {
                    "description": "提示词类型",
                    "type": "string"
                },
                "sessionId": {
                    "description": "会话id",
                    "type": "string"
                }
            }
        },
        "sessionm.SessionChatLogListReq": {
            "type": "object",
            "properties": {
                "sessionId": {
                    "type": "string"
                }
            }
        },
        "sessionm.SessionChatLogListResp": {
            "type": "object",
            "properties": {
                "chatLogs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Ai_chat_logs"
                    }
                },
                "sessionId": {
                    "type": "string"
                }
            }
        },
        "sessionm.UserSessionListReq": {
            "type": "object",
            "properties": {
                "page": {
                    "$ref": "#/definitions/dbm.Page"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "sessionm.UserSessionListResp": {
            "type": "object",
            "properties": {
                "page": {
                    "$ref": "#/definitions/dbm.Page"
                },
                "sessionList": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sessionm.UserSessionListRespData"
                    }
                }
            }
        },
        "sessionm.UserSessionListRespData": {
            "type": "object",
            "properties": {
                "creatAt": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "webapp.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码值",
                    "type": "integer"
                },
                "data": {
                    "description": "响应数据"
                },
                "message": {
                    "description": "错误消息",
                    "type": "string"
                }
            }
        }
    }
}